Python-Server/requirements.txt
Flask
datefinder
python-interface
mysql-connector